// mods/SariaShop/src/Refs/Enums.ts
export enum Currency {
    Roubles = "5449016a4bdc2d6f028b456f",
    Dollars = "5696686a4bdc2da3298b456a",
    Euros = "569668774bdc2da2298b4568",
}